+++ tests/mkDotSpago/mkDotSpago.nix
@@ -1,0 +1,5 @@
+{ buildDotSpago }:
+buildDotSpago {
+  src = ./.;
+  spagoLock = ./spago.lock;
+}